refactor: extract maven plugin core, and move element-template-c7
