+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/Camunda7ElementTemplateGenerator.java
@@ -1,0 +1,119 @@
+package io.miragon.miranum.connect.adapter.in.c7.elementtemplates;
+
+import io.miragon.miranum.connect.c7.elementtemplates.gen.Binding;
+import io.miragon.miranum.connect.c7.elementtemplates.gen.CamundaC7ElementTemplate;
+import io.miragon.miranum.connect.c7.elementtemplates.gen.Constraints;
+import io.miragon.miranum.connect.c7.elementtemplates.gen.Property;
+import io.miragon.miranum.connect.elementtemplate.api.BPMNElementType;
+import io.miragon.miranum.connect.elementtemplate.api.ElementTemplateProperty;
+import io.miragon.miranum.connect.elementtemplate.api.PropertyType;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateGenerationResult;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateGenerator;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateInfo;
+import lombok.extern.java.Log;
+
+import java.util.Arrays;
+import java.util.Objects;
+
+@Log
+public class Camunda7ElementTemplateGenerator implements ElementTemplateGenerator {
+
+    @Override
+    public ElementTemplateGenerationResult generate(ElementTemplateInfo elementTemplateInfo) {
+        var elementTemplate = new CamundaC7ElementTemplate()
+                .withName(elementTemplateInfo.getName())
+                .withId(elementTemplateInfo.getId())
+                .withAppliesTo(Arrays.stream(elementTemplateInfo.getAppliesTo()).map(BPMNElementType::getValue).toList());
+
+        // Add external task property
+        var implementationProperty = new Property()
+                .withLabel("Implementation Type")
+                .withType(PropertyType.STRING.getType())
+                .withValue("external")
+                .withEditable(false)
+                // We set the choices to null, because we don't want to have an empty
+                // choices property in the JSON. This can be fixed by not creating an
+                // empty list in the first place, but this is what got generated by the
+                // jsonschema2pojo generator.
+                .withChoices(null)
+                .withBinding(new Binding()
+                        .withType(Binding.Type.PROPERTY)
+                        .withName("camunda:type"));
+        elementTemplate.getProperties().add(implementationProperty);
+
+        // Add property for the topic of the external task
+        var implementationTopicProperty = new Property()
+                .withLabel("Topic")
+                .withType(PropertyType.STRING.getType())
+                .withValue(elementTemplateInfo.getType())
+                .withEditable(false)
+                .withChoices(null)
+                .withBinding(new Binding()
+                        .withType(Binding.Type.PROPERTY)
+                        .withName("camunda:topic"));
+        elementTemplate.getProperties().add(implementationTopicProperty);
+
+        // Add properties for input parameters
+        if (!Objects.isNull(elementTemplateInfo.getInputType())) {
+            for (var field : elementTemplateInfo.getInputType().getDeclaredFields()) {
+                var type = PropertyType.getType(field.getType());
+                var annotation = field.getAnnotation(ElementTemplateProperty.class);
+                var property = createPropertyWithPossibleAnnotation(
+                        field.getName(),
+                        type,
+                        "",
+                        annotation);
+
+                var binding = new Binding()
+                        .withType(Binding.Type.CAMUNDA_INPUT_PARAMETER)
+                        .withName(field.getName());
+
+                property.setBinding(binding);
+                elementTemplate.getProperties().add(property);
+            }
+        }
+
+        // Add properties for output parameters
+        if (!Objects.isNull(elementTemplateInfo.getOutputType())) {
+            for (var field : elementTemplateInfo.getOutputType().getDeclaredFields()) {
+                var type = PropertyType.getType(field.getType());
+                var annotation = field.getAnnotation(ElementTemplateProperty.class);
+                var property = createPropertyWithPossibleAnnotation(
+                        field.getName(),
+                        type,
+                        field.getName() + "Result",
+                        annotation);
+
+                var binding = new Binding()
+                        .withType(Binding.Type.CAMUNDA_OUTPUT_PARAMETER)
+                        .withSource("${" + field.getName() + "}");
+
+                property.setBinding(binding);
+                elementTemplate.getProperties().add(property);
+            }
+        }
+
+        var json = CamundaC7ElementTemplateConverter.toJsonString(elementTemplate);
+        return new ElementTemplateGenerationResult(elementTemplateInfo.getId(), elementTemplateInfo.getVersion(), json);
+    }
+
+    private Property createPropertyWithPossibleAnnotation(String label, PropertyType type, String value, ElementTemplateProperty propertyAnnotation) {
+        var property = new Property()
+                .withLabel(label)
+                .withType(type.getType())
+                .withChoices(null)
+                .withValue(value);
+
+        if (!Objects.isNull(propertyAnnotation)) {
+            property.setLabel(propertyAnnotation.label().isEmpty() ? label : propertyAnnotation.label());
+            property.setType(Objects.isNull(propertyAnnotation.type()) ? type.getType() : propertyAnnotation.type().getType());
+            property.setEditable(propertyAnnotation.editable());
+
+            var constraints = new Constraints();
+            constraints.setNotEmpty(propertyAnnotation.notEmpty());
+            property.setConstraints(constraints);
+        }
+
+        return property;
+    }
+}
