+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/SchemaMixin.java
