+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-c7/src/main/resources/schema/camunda-c7-element-template.json
