+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-c7/src/test/resources/expected-test-element-template.json
