+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin-core/src/main/java/io/miragon/miranum/connect/elementtemplate/core/ElementTemplateGenerationResult.java
@@ -1,0 +1,13 @@
+package io.miragon.miranum.connect.elementtemplate.core;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ElementTemplateGenerationResult {
+
+    private String name;
+    private String version;
+    private String json;
+}
