+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin-core/src/main/java/io/miragon/miranum/connect/elementtemplate/core/ElementTemplateGenerator.java
@@ -1,0 +1,6 @@
+package io.miragon.miranum.connect.elementtemplate.core;
+
+public interface ElementTemplateGenerator {
+
+    ElementTemplateGenerationResult generate(ElementTemplateInfo elementTemplateInfo);
+}
