+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin-core/src/main/java/io/miragon/miranum/connect/elementtemplate/core/ElementTemplateInfo.java
@@ -1,0 +1,24 @@
+package io.miragon.miranum.connect.elementtemplate.core;
+
+import io.miragon.miranum.connect.elementtemplate.api.BPMNElementType;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ElementTemplateInfo {
+
+    private final String name;
+
+    private final String id;
+
+    private final String version;
+
+    private final String type;
+
+    private final BPMNElementType[] appliesTo;
+
+    private final Class<?> inputType;
+
+    private final Class<?> outputType;
+}
