+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin-core/src/main/java/io/miragon/miranum/connect/elementtemplate/core/TargetPlatform.java
@@ -1,0 +1,5 @@
+package io.miragon.miranum.connect.elementtemplate.core;
+
+public enum TargetPlatform {
+    camunda7, camunda8
+}
