+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin/src/main/java/io/miragon/miranum/connect/elementtemplate/ElementTemplateGeneratorFactory.java
@@ -1,7 +1,8 @@
 package io.miragon.miranum.connect.elementtemplate;
 
 import io.miragon.miranum.connect.adapter.in.c7.elementtemplates.Camunda7ElementTemplateGenerator;
-import io.miragon.miranum.connect.elementtemplate.api.ElementTemplateGenerator;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateGenerator;
+import io.miragon.miranum.connect.elementtemplate.core.TargetPlatform;
 
 public class ElementTemplateGeneratorFactory {
