+++ connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin/src/main/java/io/miragon/miranum/connect/elementtemplate/ElementTemplateGeneratorMojo.java
@@ -1,7 +1,9 @@
 package io.miragon.miranum.connect.elementtemplate;
 
-import io.miragon.miranum.connect.elementtemplate.api.ElementTemplateGenerationResult;
 import io.miragon.miranum.connect.elementtemplate.api.GenerateElementTemplate;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateGenerationResult;
+import io.miragon.miranum.connect.elementtemplate.core.ElementTemplateInfoMapper;
+import io.miragon.miranum.connect.elementtemplate.core.TargetPlatform;
 import org.apache.maven.artifact.DependencyResolutionRequiredException;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
