--- connect/connect-core/connect-element-templates-maven-plugin/element-templates-generator-maven-plugin/src/main/java/io/miragon/miranum/connect/elementtemplate/TargetPlatform.java
@@ -1,5 +1,0 @@
-package io.miragon.miranum.connect.elementtemplate;
-
-public enum TargetPlatform {
-    camunda7, camunda8
-}
