--- connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/api/ElementTemplateGenerationResult.java
@@ -1,13 +1,0 @@
-package io.miragon.miranum.connect.elementtemplate.api;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ElementTemplateGenerationResult {
-
-    private String name;
-    private String version;
-    private String json;
-}
