--- connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/api/ElementTemplateGenerator.java
@@ -1,6 +1,0 @@
-package io.miragon.miranum.connect.elementtemplate.api;
-
-public interface ElementTemplateGenerator {
-
-    ElementTemplateGenerationResult generate(ElementTemplateInfo elementTemplateInfo);
-}
