--- connect/connect-core/connect-element-templates/element-template-api/src/main/java/io/miragon/miranum/connect/elementtemplate/api/ElementTemplateInfo.java
@@ -1,23 +1,0 @@
-package io.miragon.miranum.connect.elementtemplate.api;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ElementTemplateInfo {
-
-    private final String name;
-
-    private final String id;
-
-    private final String version;
-
-    private final String type;
-
-    private final BPMNElementType[] appliesTo;
-
-    private final Class<?> inputType;
-
-    private final Class<?> outputType;
-}
